Move notification handling to Callback class
